backend/service/reportsvc/pyproject.toml
[tool.poetry]
name = "reportsvc"
version = "0.1.0"
description = "A service for managing reports"
authors = []

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
fastapi = "0.109.2"
uvicorn = "0.22.0"
gdown = "5.1.0"
openapi-python-client = "0.19.0"
PyYAML = "6.0.1"

#[tool.poetry.group.local.dependencies]
#whitebox = { path = "../../lib/whitebox" }
#param_lib = { path = "../../lib/param_lib" }
#model_lib = { path = "../../lib/model_lib", develop = true }
#od_client = { path = "../../client/od_client", develop = true }
#av_client = { path = "../../client/av_client", develop = true }
#mac_client = { path = "../../client/mac_client", develop = true }
#db = { path = "../../db", develop = true }

#[tool.poetry.group.local]
#optional = true

[tool.poetry.scripts]
generate-client = "scripts.generate_client:main"

[build-system]
requires = ["poetry-core>=1.8.2"]
build-backend = "poetry.core.masonry.api"
